ref: extract category management from LibraryViewModel to isolated UseCase
💡 What: Extracted category modification logic (add new, update, sort ascending, etc.) out of `LibraryViewModel` and `MangaViewModel` and isolated it into a `ModifyCategoryUseCase` within a `CategoryUseCases` wrapper.
🎯 Why: Follows the principle of high cohesion and low coupling by extracting generic domain manipulation out of view models into Use Cases, reducing God Classes (`LibraryViewModel`, `MangaViewModel`) size, standardizing cross-view-model behavior and removing code duplication.
📍 From/To: Logic from `app/src/main/java/eu/kanade/tachiyomi/ui/library/LibraryViewModel.kt` and `app/src/main/java/eu/kanade/tachiyomi/ui/manga/MangaViewModel.kt` has been cleanly extracted into `app/src/main/java/org/nekomanga/usecases/category/ModifyCategoryUseCase.kt`.
